deleted unused vectorinfo dynamic api
