+++ src/lib/helper/generateMetadataJson.ts
@@ -9,7 +9,7 @@ import type { TileJson, VectorLayerTileStatLayer, VectorTileMetadata } from '$li
  */
 export const generateMetadataJson = async (tilejson: TileJson, origin: string) => {
   const pbfPath = tilejson.tiles[0].replace('{z}/{x}/{y}', '0/0/0')
-  const vectorinfoUrl = `${origin}/vectorinfo/static?path=${pbfPath}&layer_name=${tilejson.name}`
+  const vectorinfoUrl = `${origin}/vectorinfo?path=${pbfPath}&layer_name=${tilejson.name}`
   const res = await fetch(vectorinfoUrl)
   if (!res.ok) {
     throw error(res.status, { message: res.statusText })
